pyproject.toml
[tool.poetry]
name = "fuzzy-trust"
version = "0.1.0"
description = "A reputation system based on fuzzy logic"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
numpy = "^1.26.4"
ipykernel = "^6.29.5"
seaborn = "^0.13.2"
hydra-joblib-launcher = "^1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
